Many changes to avoid all compiler warnings on 32 and 64 bit Linux and Windows
git-svn-id: https://subversion.xor.aps.anl.gov/synApps/areaDetector/trunk@15490 dc6c5ff5-0b8b-c028-a01f-ffb33f00fc8b

+++ ADApp/nexusSrc/nxdataset.c
@@ -81,7 +81,7 @@ pNXDS createTextNXDataset(char *name){
   pNew->rank = 1;
   pNew->type = NX_CHAR;
   pNew->magic = MAGIC;
-  pNew->dim[0] = strlen(name);
+  pNew->dim[0] = (int)strlen(name);
   return pNew;
 }
 /*-----------------------------------------------------------------------*/
